+++ packages/network/src/json-rpc/middleware/cached-provider.ts
@@ -39,7 +39,6 @@ export class CachedProvider implements JsonRpcMiddlewareHandler {
         if (this.cachableJsonRpcMethods.includes(request.method)) {
           if (response && response.result) {
             const key = this.cacheKey(request.method, request.params!, chainId || this.defaultChainId)
-            console.debug("cache", key, response.result)
             this.setCacheValue(key, response.result)
           }
         }