+++ packages/simulator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @0xsodium/simulator
 
+## 1.15.0
+
+### Minor Changes
+
+- Support dynamic iframe size
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/transactions@1.15.0
+
 ## 1.14.0
 
 ### Minor Changes